Move exposure site scrapers into single folder
